refactor: solve circular imports Use blueprints and current_app
