+++ csc/questions/routes.py
@@ -1,83 +1,21 @@
-from __main__ import app
-from ast import literal_eval
-import json
-import logging
 from nanoid import generate
-import openai
-from quart import request, jsonify
-from main import MODEL
+from quart import Blueprint, request, jsonify
 import prompts.prompts as prompts
-from database import check_db, get_all_questions, add_question_db, get_db, update_question_db, delete_question_db
-from user.routes import get_user_info
+from database import check_db, get_db
+from utils.user import get_user_info
 from utils.utils import checkResponseSuccess, prettify_questions
 
-async def openai_generate_qns_add_db(room_id, messages):
-    logging.info("{room_id}: Querying OpenAI")
-    response = openai.ChatCompletion.create(
-        model=MODEL,
-        messages=messages,
-        temperature=0.7,
-    )
-    questions = response['choices'][0]['message']['content']
-
-    # # add questions in the form of room-id to questions key value pairs
-    logging.info(f"{room_id}: Adding questions to database: {questions}")
-
-    question_arr = parse_questions(questions)
-    
-    await add_questions_to_room_collection(question_arr, room_id, "csc")
-
-    logging.info(f"{room_id}: Done adding questions to database")
-
-def parse_questions(questions):
-    if questions[0] == "[" and questions[-1] == "]":
-        # This may be dangerous in the case of prompt injection
-        return literal_eval(questions)
-    else:
-        return [questions]
-
-async def add_questions_to_room_collection(questions, room_id, game_type):
-    try:
-        await get_db()["Rooms"].insert_one({
-            '_id': room_id,
-            'game_type': game_type,
-            'is_published': False,
-            'questions': questions
-        })
-    except Exception as e:
-        logging.error(f"Error: {str(e)}")
-        return jsonify({"message": f"Error: adding questions to collection"}), 500
-
-async def generate_unique_room_id():
-    while True:
-        room_id = generate(size=6)
-        room = await get_db()["Rooms"].find_one({'_id': room_id})
-        if not room:
-            logging.info(f"Unique room {room_id} found")
-            break
-    return room_id
-
-async def set_room_published_status(room_id, set_is_published):
-    existing_room = await get_db()["Rooms"].find_one({"_id": room_id})
-    if not existing_room:
-        return {"error": "Room not found"}, 404
-
-    await get_db()["Rooms"].update_one({"_id": room_id},
-                                        {'$set': {
-                                        'is_published': set_is_published
-                                        }})
-    
-    return {"message": f"Room {room_id} {'published' if set_is_published else 'unpublished'} successfully"}
+csc_questions_bp = Blueprint('csc_questions_bp', __name__, url_prefix='/csc/questions')
 
 
 # FOR TESTING PURPOSES ONLY
 # Define a route to check if the database is accessible
-@app.route('/check_database', methods=['GET'])
+@csc_questions_bp.route('/check_database', methods=['GET'])
 async def check_database():
     res = await check_db()
     return res
 
-@app.route('/csc/questions', methods=['GET'])
+@csc_questions_bp.route('/', methods=['GET'])
 async def get_csc_questions():
     # TODO: check if this works
     user_info = get_user_info()
@@ -85,6 +23,7 @@ async def get_csc_questions():
         return user_info # will contain error and status message
 
     user_email = user_info[0].get("email")
+    # user_email = "user@example.com"
 
     user = await get_db()['Users'].find_one({"_id": user_email})
     if not user:
@@ -95,7 +34,7 @@ async def get_csc_questions():
     return prettify_questions(questions), 200
 
 
-@app.route('/csc/questions/<id>', methods=['PUT'])
+@csc_questions_bp.route('/<id>', methods=['PUT'])
 async def update_csc_question(id):
     request_data = await request.json
     content = request_data.get('content')
@@ -126,7 +65,7 @@ async def update_csc_question(id):
         error_message = f"Error: {str(e)}"
         return jsonify({"message": error_message}), 500
 
-@app.route('/csc/questions/create', methods=['POST'])
+@csc_questions_bp.route('/csc/questions/create', methods=['POST'])
 async def create_csc_question():
     # TODO: check if this works
     user_info = get_user_info()
@@ -160,7 +99,7 @@ def generate_unique_question_id(questions):
         question_id = generate(size=3)
     return question_id
 
-@app.route('/csc/questions/<id>', methods=['DELETE'])
+@csc_questions_bp.route('/<id>', methods=['DELETE'])
 async def delete_csc_question(id):
     # TODO: check if this works
     user_info = get_user_info()
